+++ NEWS.md
@@ -5,6 +5,7 @@
 * fixed a bug in germancredit data set
 * scorecard2 supports to adjust oversampling 
 * provides replace_na function to replace missing values
+* a new function var_scale is provided for variable scaling 
 
 # scorecard 0.2.7